simplify trainer by moving common utils to utils.py
